Makefile changes
LatexMk on TeX Live on Windows breaks the pdflatex build if the argument
containing the path to the tex build to build contains slashes. On Windows,
forward slashes are changed to backslashes, which are not escaped when passed
on to pdflatex, so pdflatex thinks the file path is an unrecognized command...

To try and circumvent this, I've changed the Makefile so we always cd into the
folder where the test tex is located, and then specify the path to the tex file
as just a file name (without any slashes).

Also, I've updated the Makefile to try and be smarter about detecting python.