AppVeyor: try to fix hanging TL2017 install
Working theory is that the install succeeded, but is hanging on a
"Press any key to continue . . ." prompt, waiting for user input, because
this is what happens when running `.\install-tl-advanced.bat -no-gui --help`.

Try to pipe an ENTER key to the command to circumvent this (it works when
running with --help, at least).