ltxpect: add E2E verification tests for test engine
Motivation is that the comparison tests are no longer working (tests that
should have failed are currently passing), due to a regression related to a
new version of ImageMagick that (yet again) changes the output of the compare
utility.

The regression will be addressed in a later changeset, but before fixing it, I
wanted a new test suite that would catch this so it doesn't happen again.

The key idea behind this new test suite is to have a small set of very simple
LaTeX test documents (should only require a single run using pdflatex to compile),
which we want to pre-compile to produce a PDF, such that we can use invoke the test
engine with that PDF as a proto PDF, effectively comparing the test against itself.
When comparing against itself, we expect the test engine to produce a success result
with no diff. Then, we also have the concept of a "skew" which we can apply to each
test case, to make the test produce a _slightly_ different output, which we again
can use as a proto PDF, and this time we expect to get a negative test result with
a diff.