ltxpect: fix ImageMagick regression
The output of the compare utility (for the ae metric) seems to have "flipped"
(i.e. changed from 'AAA (BBB)' to 'BBB (AAA)').

I suspect this change: https://github.com/ImageMagick/ImageMagick/commit/8a735ac1aaacb1a5c0ce597d9cacbe7a6e73a741

but the code is quite hard to read. In case it will be useful for future
re-visits, these code locations were of interest:

- https://github.com/ImageMagick/ImageMagick/blob/7404bd00879aa0a515d7760d6fdd889e39542f41/MagickWand/compare.c#L1321-L1330
- https://github.com/ImageMagick/ImageMagick/blob/7404bd00879aa0a515d7760d6fdd889e39542f41/MagickWand/compare.c#L1252
- https://github.com/ImageMagick/ImageMagick/blob/7404bd00879aa0a515d7760d6fdd889e39542f41/MagickCore/compare.c#L2294
- https://github.com/ImageMagick/ImageMagick/blob/7404bd00879aa0a515d7760d6fdd889e39542f41/MagickCore/compare.c#L307

The entry point for the compare utility is in MagickWand, but a lot of the functionality is in MagickCore.